+++ easy_mode/changelog.md
@@ -1,6 +1,6 @@
 ﻿# 0.0.2
 ## Features
-- Added option for automatic exp magnet
+- Added a lot of settings and hotkeys
 
 # 0.0.1
 - Initial release